day-2/python-drills/helloworld_nikita/pyproject.toml
                                                      [project]
name = "helloworld_nikita"
version = "0.1.0"
description = "Hello world CLI and library"
authors = [{name = "Nikita", email = "nikitaisrani08@gmail.com"}]
dependencies = ["typer", "pyyaml"]


[project.scripts]
hello = "helloworld_nikita.main:app"

[build-system]
requires = ["setuptools>=61.0", "wheel"]
build-backend = "setuptools.build_meta"

[tool.setuptools.package-data]
"helloworld_nikita" = ["_default_config.yaml"]
